fix(hidden-games): finish the rebase onto master
The branch forked at DB v20 while master moved to v30, so the hidden_games and appType migrations are renumbered to 30->31 and 31->32 with the v32 schema export, and the deep-history migration chain is extended to match. Daily-progress exclusion moves into the committer/writer path master extracted, and the visibility effect adopts master's Long XP. None of this changes the feature's specified behavior.

+++ app/src/main/java/com/example/backlogium/domain/GameVisibilityUseCase.kt
@@ -20,8 +20,8 @@ data class VisibilityChangeEffect(
     val names: List<String>,
     /** True for a hide, false for an unhide; the disclosure reads differently for each. */
     val hiding: Boolean,
-    val totalXpBefore: Int,
-    val totalXpAfter: Int,
+    val totalXpBefore: Long,
+    val totalXpAfter: Long,
     val levelBefore: Int,
     val levelAfter: Int,
     /** Names of games whose goal designation this change will clear. Empty when none applies. */
